==> 1. Data Import and Descriptive Stats.ipynb <==
### Contents
- Import
- Descriptive stats

# 01. Importing Libraries

# Import libraries
import pandas as pd
import numpy as np
import os

# 02. Importing Data

df = pd.read_csv(r"C:\Users\Kieran\Documents\Career Foundry\Data Immersion\4.0 Python\Instacart Basket Analysis\02 Data\Original Data\orders.csv", index_col=False)

#Setting up folder path
path = r"C:\Users\Kieran\Documents\Career Foundry\Data Immersion\4.0 Python\Instacart Basket Analysis"

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col = False)

df.head()

df.tail()

df.shape

df.info()

df.dtypes

df.columns

df.describe()

# list of columns needed
vars_list = ['order_id', 'user_id', 'order_number', 'order_dow', 'order_hour_of_day', 'days_since_prior_order']

#importing dataset only columns needed (not eval_set)
df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), usecols = vars_list)

df.head()

# Importing products.csv

#Q7
df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col = False)

#Q8
df_prods.head(20)

df_prods.tail(35)

#Q9
df_prods.columns

#Q10
df_prods.shape

df_prods.describe()

Q11a) The maximum value of aisle_id is 134

df_prods.dtypes

Q11b) The data type of department_id is integer.
