==> 2. Data Wrangling.ipynb <==
### Contents
- Droped eval_set column
- Changed column names
- Changed data types
- Created department data dictionary
- Subsetting
- Export

import pandas as pd
import numpy
import os

path = r"C:\Users\Kieran\Documents\Career Foundry\Data Immersion\4.0 Python\Instacart Basket Analysis"

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'orders.csv'), index_col=False)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col=False)

Dropping columns, renaming columns, changing data types, transposing data are examples of data wrangling

#Dropping eval_set column from orders. This does not change the df_ords df. 
df_ords.drop(columns = ['eval_set'])

# value_counts() counts all the values 
df_ords['days_since_prior_order'].value_counts(dropna = False) # including dropna=False insures NaN values are counted

#change column name
df_ords.rename(columns = {'order_dow' : 'orders_day_of_the_week'}, inplace = True)#inplace=True means we overwrite original name

df_ords.head()

#changing data type. You would do this for variables like order_id which do not need statistical analysis as they are keys.
df_ords['order_id'] = df_ords['order_id'].astype('str') #we use = because we are overwriting existing order_id column

df_ords['order_id'].dtype #'o' stands for object, python version of a string

df_dep = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'departments.csv'), index_col = False)

df_dep.head()

# Transposing df_dep
df_dep.T

#new df tranposed
df_dep_t = df_dep.T

df_dep_t

# adding an index
df_dep_t.reset_index()

# take the first row for header
new_header = df_dep_t.iloc[0]

new_header

# take the data from under the header row as a new df
df_dep_t_new = df_dep_t[1:]

df_dep_t_new

# set the header row as a the df header
df_dep_t_new.columns = new_header

df_dep_t_new

# transforming df in to a data dictionary
data_dict = df_dep_t_new.to_dict('index') # 'index' tells python to use numbered rows as key values

data_dict

df_prods.head()

# use data dictionary to determine what department_id stand for
print(data_dict.get('19'))

# Subsetting

Subsetting refers to creating a smaller data set from a whole data set based on a filter.

# indexing code alone. Just searches for the data in question by going through every row and evaluating against == 19. 
df_prods['department_id'] == 19

# wrapping indexing code in df_prods[] tells python to map the true/false output from the index code, only showing true values
df_prods[df_prods['department_id'] == 19]

# subset from df_prods df that contains data from the snacks department
df_snacks = df_prods[df_prods['department_id'] == 19]

df_snacks.head()

# using loc function to subset. loc looks for column name. 
df_snacks_2 = df_prods.loc[df_prods['department_id'] == 19]

# could use isin instead of =. this is useful for checking multiple values
df_snacks_3 = df_prods.loc[df_prods['department_id'].isin([19])]

# 4.4 Questions

df_ords.head()

# df_ords still have eval_set so I will remove and overwrite
df_ords = df_ords.drop(columns = ['eval_set'])

df_ords.dtypes

# Q2 user_id does not need to be integer 
df_ords['user_id'] = df_ords['user_id'].astype('str')
df_ords.dtypes

df_ords.columns = ['order_id', 'user_id', 'order_number', 'orders_day_of_the_week', 'order_hour_of_day', 'days_since_last_order']

# Q3 changing column name
df_ords.rename(columns = {'days_since_prior_order' : 'days_since_last_order'})

# Q4. 10 is the busiest hour for placing orders.
df_ords['order_hour_of_day'].value_counts(dropna = False)

# Q5
print(data_dict.get('4'))

# Q6
df_breakfast = df_prods[df_prods['department_id'] == 14]

df_breakfast.head()

# Q7 find observations from certain dpeartments
data_dict

# required department_id's
party = '5', '7', '12', '20'

# party subset
df_party = df_prods[df_prods['department_id'].isin(party)]

df_party.head()

# Q8 
df_party.shape

# Q9
user_id_1 = df_ords.loc[df_ords['user_id'] == '1']

user_id_1.head()

# Q10
user_id_1.describe()

Exporting dataframes

# orders wrangled csv
df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangles.csv'))

# departments wrangled
df_dep_t_new.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'Departments_wrangled.csv'))
