==> 3. Data Consitency Checks.ipynb <==
### Contents  
- Check mixed data types
- Check missing values
- Check duplicates

import pandas as pd
import numpy
import os

path = r"C:\Users\Kieran\Documents\Career Foundry\Data Immersion\4.0 Python\Instacart Basket Analysis"

df_ords = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_wrangled.csv'), index_col=False)

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'products.csv'), index_col=False)

df_test = pd.DataFrame()

df_test['mix'] = ['a', 'b', 1, True] 

df_test

# function to check for mixed data type columns
for col in df_test.columns.tolist():
  weird = (df_test[[col]].applymap(type) != df_test[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_test[weird]) > 0:
    print (col)

# convert column data type to str
df_test['mix'] = df_test['mix'].astype('str')

# Missing Values

# finding missing values 
df_prods.isnull().sum()

# to view those values, create a subset of the df
df_nan = df_prods[df_prods['product_name'].isnull() == True]
df_nan

imputing missing values with mean:  
df['column with missings'].fillna(mean value, inplace=True)

imputing missing values with median:  
df['column with missings'].fillna(median value, inplace=True)

df_prods.shape

Because the missing values in df_prods are strings we cant impute. Instead we will remove those observations

# creating a new df without nan observations
df_prods_clean = df_prods[df_prods['product_name'].isnull() == False]

# running shape to confirm new df has less observations
df_prods_clean.shape

could overwrite the original df:  
df_prods[‘product_name’].dropna(inplace = True)

# Duplicates

Need to clarify what kind of duplicates we are looking for. Duplicates in one column may be acceptable (e.g same product_id). Full duplicates where whole rows are exactly the same are more likely to be mistakes and need removing.  

# looking for full duplicates in df
df_dups = df_prods_clean[df_prods_clean.duplicated()]
df_dups

df_prods_clean.shape

# new df that does not contain duplicates
df_prods_clean_no_dups = df_prods_clean.drop_duplicates()
df_prods_clean_no_dups.shape

# exporting cleaned df as csv
df_prods_clean_no_dups.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'))

# 4.5 Questions

# Q2 
df_prods.describe()

# checking that the max price = 99999 observation has been removed in the clean df
df_prods_clean_no_dups.describe()

# the anomaly is still in the data. I wont remove until instructed.
df_prods_clean_no_dups[df_prods_clean_no_dups['prices'] == 99999]

# Q3 checking for mixed data types in df_ords
for col in df_ords.columns.tolist():
  weird = (df_ords[[col]].applymap(type) != df_ords[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df_ords[weird]) > 0:
    print (col)

No mixed data types in df_ords. 

# Q5
df_ords.isnull().sum()

# viewing null values
df_ords_nan = df_ords[df_ords['days_since_last_order'].isnull() == True]
df_ords_nan

df_ords.shape

Q5/Q6 The missing values are all in days_since_last_order column. This suggests that the order is this customers first order. An appropiate imputed value would therefore be 0.

df_ords['days_since_last_order'].fillna(0, inplace=True)

df_ords.isnull().sum()

# Q7
df_ords_dup = df_ords[df_ords.duplicated()]
df_ords_dup

No full duplicates

# checking if there are duplicate order id
df_ords_dup_ord = df_ords[df_ords['order_id'].duplicated()]
df_ords_dup_ord

No duplicates in order_id

# Q9 Export df_ords
df_ords.to_csv(os.path.join(path, '02 Data', 'Prepared Data', 'orders_checked.csv'))
