==> 4. Combining Orders & Products.ipynb <==
### Contents  
- Combined orders and products

import pandas as pd
import numpy as np
import os

# Q3 Import order_products_combined

path = r"C:\Users\Kieran\Documents\Career Foundry\Data Immersion\4.0 Python\Instacart Basket Analysis"

df_ords = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_combined.pkl'))

df_ords.head()

# Q4 shape of df is the same

df_ords.shape

Import produdcts data

df_prods = pd.read_csv(os.path.join(path, '02 Data', 'Prepared Data', 'products_checked.csv'), index_col = False)

df_prods.head()

I will delete the uneccesary unamed column

df_prods.drop(df_prods.columns[0], axis=1, inplace=True)

df_prods.head()

df_prods.shape

# Q5 Combine datasets using merge on product_id

df_merged = df_ords.merge(df_prods, on = 'product_id')

df_merged.head()

# Q6 confirm merge 

df_merged['_merge'].value_counts()

# Q7 Export merged file 

df_merged.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))
