==> 5. Deriving New Variables.ipynb <==
## Contents
- Price label function
- Busiest day function
- Busiest hour function

import pandas as pd
import numpy as np
import os

path = r"C:\Users\Kieran\Documents\Career Foundry\Data Immersion\4.0 Python\Instacart Basket Analysis"

# Import data

ords_prods = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'orders_products_merged.pkl'))

# working we subset of data, first million rows
df = ords_prods[:1000000]

# Defining price label function

# defining price_label function
def price_label(row):
    if row['prices'] <= 5:
        return 'Low range product'
    elif (row['prices'] > 5) and (row['prices'] <= 15):
        return 'Mid range product'
    elif row['prices'] > 15:
        return 'High range product'
    else: return np.nan

# applying function to df
df['price_range'] = df.apply(price_label, axis=1)

df['price_range'].value_counts(dropna = False)

df['prices'].max()

# If statements with loc()
Using loc() applies the conditional logic of if statements without explicitly creating if-else statements.

df.loc[df['prices'] > 15, 'price_range_loc'] = 'High-range product'

df.loc[(df['prices'] <= 15) & (df['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

df.loc[df['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

df['price_range_loc'].value_counts(dropna = False)

# Using loc() on entire dataframe

ords_prods.loc[ords_prods['prices'] > 15, 'price_range_loc'] = 'High-range product'

ords_prods.loc[(ords_prods['prices'] <= 15) & (ords_prods['prices'] > 5), 'price_range_loc'] = 'Mid-range product'

ords_prods.loc[ords_prods['prices'] <= 5, 'price_range_loc'] = 'Low-range product'

ords_prods['price_range_loc'].value_counts(dropna = False)

# If Statements with For-Loops

Busiest day:

ords_prods['orders_day_of_the_week'].value_counts(dropna = False) # 0 = Sunday

New column 'busiest day' created using a for loop

# empty list
result = []

# for-loop
for value in ords_prods['orders_day_of_the_week']:
    if value == 0:
        result.append('Busiest day')
    elif value == 4:
        result.append('Least busy')
    else:
        result.append('Regularly busy')

result

# creating new column in ords_prods and filling with result list
ords_prods['busiest day'] = result

ords_prods['busiest day'].value_counts(dropna = False)

# Exercise

Q2

# empty list
busiest_result = []

#for loop busiest days
for x in ords_prods['orders_day_of_the_week']:
    if x == 0 or x == 1:
        busiest_result.append('Busiest days')
    elif x == 4 or x == 3:
        busiest_result.append('Slowest days')
    else:
        busiest_result.append('Regularly busy')

# check list
busiest_result

#add column
ords_prods['busiest_days'] = busiest_result

Q3

#check values
ords_prods['busiest_days'].value_counts(dropna = False)

Both the busiest and least busy days have increased in number compared to when I was only counting one day for each. This makes sense as now two days count as 'busy' and two days count as 'slow' or least busy. 

Q4 

#checking column names
ords_prods.columns

#frequency of order hours
ords_prods['order_hour_of_day'].value_counts(dropna = False)

# there are 24 hours, so 8 hours in each period. Empty list:
hours = []

for x in ords_prods['order_hour_of_day']:
    if x in range(9, 16):
        hours.append('Most orders')
    elif x == 7 or x == 8 or x in range(17, 22):
        hours.append('Average orders')
    else:
        hours.append('Fewest orders')

#add column
ords_prods['busiest_period_of_day'] = hours

Q5

#frequency of column
ords_prods['busiest_period_of_day'].value_counts(dropna = False)

export to pickle

ords_prods.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_new_variables.pkl'))
