==> 6. Loyalty flag.ipynb <==
## Contents
 - Reducing memory usage
 - Creating loyalty flag
 - Descriptive stats for loyal customers
 - Descriptive stats for regular customers
 - Descriptive stats for new customers
 - Create spending flag
 - Frequent vs non-frequent customers

import pandas as pd
import numpy as np
import os

path = r"C:\Users\Kieran\Documents\Career Foundry\Data Immersion\4.0 Python\Instacart Basket Analysis"

# Import data
df = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_new_variables.pkl'))

Checking data types to see if I can change and reduce the memory needed:

df.dtypes

orders_day_of_the_week, order_hour_of_the_day, department_id can be changed to int16 which will reduce memory usage

df.orders_day_of_the_week = df.orders_day_of_the_week.astype('uint16')

df.order_hour_of_day = df.order_hour_of_day.astype('uint16')

df.department_id = df.department_id.astype('uint16')

## Q2

df.groupby('department_id').agg({'order_number': ['mean']})

## Q3  
The results here contain all of the department id's, and the mean values are different compared to when I only completed this action on a subset of the data. 

## Q4  
Creating a loyalty flag:

# max_order column
df['max_order'] = df.groupby(['user_id'])['order_number'].transform(np.max)

df.head()

creating loyalty_flag column

df.loc[df['max_order'] <= 10, 'loyalty_flag'] = 'New Customer' 

df.loc[(df['max_order'] <= 40) & (df['max_order'] > 10), 'loyalty_flag'] = 'Regular Customer'

df.loc[df['max_order'] > 40, 'loyalty_flag'] = 'Loyal Customer'

# Checking values
df.loyalty_flag.value_counts(dropna = False)

# changing dtype for loyalty_flag to keep memory usage small
df.loyalty_flag = df.loyalty_flag.astype('category')

df.prices.max()

# changing prices dtype
df.prices = df.prices.astype('int32')

## Q5  
Spending habits of customers with different loyalty

# df with just flags and prices columns
df_prices_flags = df[['prices', 'loyalty_flag']].copy() 

df_prices_flags.head()

# Creating df for loyal cusomer only to get descriptive stats

# creating df for loyal customers only
df_loyal = df_prices_flags[df_prices_flags['loyalty_flag'] == 'Loyal Customer']

df_loyal.head()

df_loyal.shape

df_loyal['prices'].mean()

df_loyal['prices'].max()

df_loyal['prices'].min()

df_loyal['prices'].median()

# Regular customer stats

df_regular = df_prices_flags[df_prices_flags['loyalty_flag'] == 'Regular Customer']

df_regular.head()

df_regular.shape

df_regular['prices'].mean()

df_regular['prices'].min()

df_regular['prices'].max()

df_regular['prices'].median()

# New customers stats

df_new = df_prices_flags[df_prices_flags['loyalty_flag'] == 'New Customer']

df_new.head()

df_new['prices'].mean()

df_new['prices'].median()

df_new['prices'].min()

df_new['prices'].max()

## Q6

Creating a spending flag:

# start with column showing mean spend grouped by user_id
df['mean_spend'] = df.groupby(['user_id'])['prices'].transform(np.mean)

df.head()

# creating spend flag - low spender
df.loc[df['mean_spend'] < 10, 'spend_flag'] = 'Low spender'

# high spender
df.loc[df['mean_spend'] >= 10, 'spend_flag'] = 'High spender'

df.head()

df.spend_flag.value_counts(dropna = False)

## Q7

Frequent vs non-frequent users

# order frequency column
df['order_freq'] = df.groupby(['user_id'])['days_since_last_order'].transform(np.median)

df.head()

# order freq flag column - non frequent
df.loc[df['order_freq'] > 20, 'order_freq_flag'] = 'Non-frequent customer' 

# regular customer
df.loc[(df['order_freq'] > 10) & (df['order_freq'] <= 20), 'order_freq_flag'] = 'Regular customer'

# frequent customer
df.loc[df['order_freq'] <= 10, 'order_freq_flag'] = 'Frequent customer'

df['order_freq_flag'].value_counts(dropna = False)

# Export data

df.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', '4.8 ords_prods.pkl'))
