==> 7. Cust wrangling and combining.ipynb <==
## Contents
Wrangling;
- renaming cust columns
- checking missing values
- checking duplicates
- checking data types  
Combined cust and ords_prods data

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns
import os
import scipy

path = r"C:\Users\Kieran\Documents\Career Foundry\Data Immersion\4.0 Python\Instacart Basket Analysis"

### Import customers data

df = pd.read_csv(os.path.join(path, '02 Data', 'Original Data', 'customers.csv'))

df.head()

df.shape

df.describe()

#### As part of wrangling I will rename columns

df = df.rename(columns = {'First Name':'first_name', 'Surnam':'surname', 'Gender':'gender', 'STATE':'state', 'Age':'age'})

df.dtypes

### Checking for missing values

df.isnull().sum()

The only missing values are first names. I am unlikely to need this data for my analysis. I will delete the first name column

df = df.drop('first_name', 1)

### Checking for duplicates

df_dup = df[df.duplicated()]
df_dup

No full duplicates.

### Checking for mixed data types

for col in df.columns.tolist():
  weird = (df[[col]].applymap(type) != df[[col]].iloc[0].apply(type)).any(axis = 1)
  if len (df[weird]) > 0:
    print (col)

No mixed data types.

### Q6 Import prepared data and combine with customer data

# importing prepared prods and ords data
df_prod = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', '4.8 ords_prods.pkl'))

df_prod.head()

### I will merge the ords and prods data and the customer data on the user_id column

# first I will delete the current _merge column in the ords and prods df, so i can tell the merge with customer df has worked
df_prod = df_prod.drop('_merge', 1)

#### I get a memory error when trying to merge df's at their current size

#checking memory size of df
df.info()

#checking memory size of df
df_prod.info()

#### I will adjust some column dtypes to lower memory usage

#checking max value to see if i can reduce memory usage without hurting analysis
df_prod['order_freq'].max()

df_prod['order_freq'] = df_prod['order_freq'].astype('uint16')

df_prod['spend_flag'] = df_prod['spend_flag'].astype('category')

df_prod['mean_spend'].max()

df_prod['mean_spend'] = df_prod['mean_spend'].astype('float16')

df_prod['max_order'].max()

df_prod['max_order'] = df_prod['max_order'].astype('int16')

df_prod[['busiest_period_of_day', 'busiest_days']] = df_prod[['busiest_period_of_day', 'busiest_days']].astype('category')

df_prod[['busiest day', 'price_range_loc']] = df_prod[['busiest day', 'price_range_loc']].astype('category')

df_prod['aisle_id'].max()

df_prod['aisle_id'] = df_prod['aisle_id'].astype('int16')

df_prod['product_name'] = df_prod['product_name'].astype('category')

df_prod['reordered'].max()

#deleting reordered column as it adds nothing to analysis
df_prod = df_prod.drop('reordered', 1)

df_prod['add_to_cart_order'].max()

df_prod['add_to_cart_order'] = df_prod['add_to_cart_order'].astype('int16')

df_prod.info()

### Making those adjustments to dtypes has reduced the memory usage from 4.9GB to 2.6GB.  
I will try to merge the df_prod and customer df again

df_merge = df_prod.merge(df, on = 'user_id', indicator = True)

df_merge.head()

df_merge['_merge'].value_counts(dropna = False)

#### Merge was successful 

### Export

df_merge.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_cust(4.9 P1).pkl'))
