==> 10. High-Low Spender.ipynb <==
# High/Low spender

### Contents
- Crosstab of spending flag and regions
- Exclusion flag for low activity customers
- New df for most active customers created

import pandas as pd
import numpy as np
import os

path = r"C:\Users\Kieran\Documents\Career Foundry\Data Immersion\4.0 Python\Instacart Basket Analysis"

df = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', '4.10 inc region.pkl'))

#### Q3 Determine whether there’s a difference in spending habits between the different U.S. regions. 

df.columns

crosstab = pd.crosstab(df['spend_flag'], df['region'])

crosstab

#### Q4 Creat exclusion flag for customers with less than 5 orders, and exclude them from data

# creating flag for low activity
df.loc[df['order_number'] < 5, 'activity_flag'] = 'low-activity'

# creating flag for regular activity
df.loc[df['order_number'] >= 5, 'activity_flag'] = 'regular_activity'

df['activity_flag'].value_counts(dropna=False)

#### New df excluding low activity customers

df_active = df[df['activity_flag'] == 'regular_activity']

df_active.info()

df_active['activity_flag'].value_counts(dropna=False)

#### Exporting df_active

df_active.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'Active Cust only.pkl'))
