==> 11. Regions.ipynb <==
### Contents
- Crosstab days since last order and number of orders
- Surname column drop
- Create regions column
- Export as df_concat

import pandas as pd
import numpy as np
import os

path = r"C:\Users\Kieran\Documents\Career Foundry\Data Immersion\4.0 Python\Instacart Basket Analysis"

# import df
df = pd.read_pickle(os.path.join(path, '02 Data', 'Prepared Data', 'ords_prods_cust(4.9 P1).pkl'))

### Crosstabs

Crosstab is like creating a pivot table.Useful for comparing columns of data

# creates crosstab betweetn days since last order and order number, to check missing values
crosstab = pd.crosstab(df['days_since_last_order'], df['order_number'], dropna=False)

crosstab

Columns are order numbers, row is days since last order. The first column (first order) is mainly zeros which makes sense. There cant be 'days since last order' in the first order column.

# copies crosstab to clipboard to look at in excel
crosstab.to_clipboard()

df.head()

df.columns

#### Dropping surname column for data protection and _merge column as it is unnecessary

df_2 = df.drop(['surname', '_merge'], axis=1)

df_2.columns

### Creating lists of regions

northeast = ['Maine', 'New Hampshire', 'Vermont', 'Massachusetts', 'Rhode Island', 'Connecticut', 'New York', 'Pennsylvania', 'New Jersey' ]

midwest = ['Wisconsin', 'Michigan', 'Illinois', 'Indiana', 'Ohio', 'North Dakota', 'South Dakota', 'Nebraska', 'Kansas', 'Minnesota', 'Iowa', 'Missouri']

south = ['Delaware', 'Maryland', 'District of Columbia', 'Virginia', 'West Virginia', 'North Carolina', 'South Carolina', 'Georgia', 'Florida', 'Kentucky', 'Tennessee', 'Mississippi', 'Alabama', 'Oklahoma', 'Texas', 'Arkansas', 'Louisiana']

west = ['Idaho', 'Montana', 'Wyoming', 'Nevada', 'Utah', 'Colorado', 'Arizona', 'New Mexico', 'Alaska', 'Washington', 'Oregon', 'California', 'Hawaii']

#### Defining function to creat region column

def region(row):
    if row['state'] in northeast:
        return 'Northeast'
    elif row['state'] in midwest:
        return 'Midwest'
    elif row['state'] in south:
        return 'South'
    elif row['state'] in west:
        return 'West'
    else: return np.nan

# df is very large, so applying region function will likely cause memory error
df_2.info()

#### I will split the df in to 3 before applying function

df_split1 = df_2[:10000000]

df_split2 = df_2[10000001:20000000]

df_split3 = df_2[20000001:]

#### Applying function to first split df

df_split1['region'] = df_split1.apply(region, axis=1)

df_split1.head()

df_split1['region'].value_counts(dropna=False)

#### Applying function to df_split2

df_split2['region'] = df_split2.apply(region, axis=1)

df_split2.head()

df_split2['region'].value_counts(dropna=False)

#### Applying function to df_split3

df_split3['region'] = df_split3.apply(region, axis=1)

df_split3.head()

df_split3['region'].value_counts(dropna=False)

### Concatenate df's

frames = [df_split1, df_split2, df_split3]

df_concat = pd.concat(frames)

# checking that concat df is same length as previous df_2
df_concat.shape

### Exporting df_concat to work on later

df_concat.info()

#### Changing region and fam_status dtype to reduce memory usage

df_concat[['region', 'fam_status']] = df_concat[['region', 'fam_status']].astype('category')

df_concat.info()

df_concat.to_pickle(os.path.join(path, '02 Data', 'Prepared Data', '4.10 inc region.pkl'))
